Module gwasm_api::prelude

The gwasm-api prelude

The purpose of this module is to alleviate imports of common structures and functions by adding a glob import to the top of the gwasm-api heavy modules:

use gwasm_api::prelude::*;

Re-exports

pub use super::task::ComputedSubtask;

pub use super::task::ComputedTask;

pub use super::task::GWasmBinary;

pub use super::task::Options;

pub use super::task::Subtask;

pub use super::task::Task;

pub use super::task::TaskBuilder;

pub use super::timeout::Timeout;

pub use super::compute;

pub use super::ProgressUpdate;

Enums

Net